nl_sql_maker/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vbg-generate"
version = "0.1.0"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
description = "A Natural Language to SQL converter with a modular pipeline and a runtime CLI."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

# Runtime + pipeline deps
dependencies = [
  "lark>=1.1.7",
  "PyYAML>=6.0",
  "shapely>=2.0.0",
  "SQLAlchemy>=2.0",   # optional for future engines; stdlib sqlite3 is used in runtime
  "pysqlite3",         # switch to 'pysqlite3-binary' if build issues, optional on many systems
  "inflect>=7.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "flake8",
]

# Console entry points
[project.scripts]
vbg_generate = "vbg_generate.vbg_pipeline:main"   # pipeline A→K
vbg_runtime  = "vbg_runtime.cli:main"             # NL→SQL runtime CLI
vbg_phrase_explorer = "vbg_tools.phrase_explorer:main"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
